+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Version 1.4.0 (TBD)
 * Added `Transformers#nest` which takes another `Transformer` and applies it to each level of a nested data structure (e.g. `Map` or `Sequence`).
+* Added `Transformers#copy` which copies the value associated with one key and also associates it with another key.
 
 #### Version 1.3.1 (December 5, 2018)
 * Fixed some bugs related to serializing and deserializing some built-in `Transformers`.